# openapi.yaml
openapi: 3.0.3
info:
  title: FitReg API
  description: |
    REST API for FitReg, a fitness coaching platform where coaches assign and manage
    workouts for their students.

    **Authentication:** All endpoints except `/health` and `POST /api/auth/google`
    require a JWT Bearer token obtained via Google OAuth.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/google:
    post:
      tags: [auth]
      summary: Exchange Google credential for JWT
      description: |
        Rate limited to 10 requests per IP per minute.
        Verifies the Google ID token, upserts the user, and returns a JWT.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [credential]
              properties:
                credential:
                  type: string
                  description: Google ID token from Google OAuth
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT to use in Authorization header
                  user:
                    $ref: "#/components/schemas/UserProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/me:
    get:
      tags: [me]
      summary: Get own profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [me]
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/me/avatar:
    post:
      tags: [me]
      summary: Upload custom avatar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
      responses:
        "200":
          description: Avatar uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [me]
      summary: Delete custom avatar
      responses:
        "200":
          description: Avatar deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/coach-request:
    get:
      tags: [coach-students]
      summary: Get current coach request status
      responses:
        "200":
          description: Request status
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [coach-students]
      summary: Request a coach (student initiates)
      responses:
        "200":
          description: Request created
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/workouts:
    get:
      tags: [workouts]
      summary: List own workouts
      responses:
        "200":
          description: List of workouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workout"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [workouts]
      summary: Create workout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkoutRequest"
      responses:
        "201":
          description: Created workout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/coach/students:
    get:
      tags: [coach-students]
      summary: List active students
      responses:
        "200":
          description: Students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoachStudentInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/coach/daily-summary:
    get:
      tags: [coach-students]
      summary: Today's workout summary for all students
      responses:
        "200":
          description: Daily summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailySummaryItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/coach/assigned-workouts:
    get:
      tags: [coach-assigned-workouts]
      summary: List all assigned workouts (coach view)
      responses:
        "200":
          description: Assigned workouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssignedWorkout"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [coach-assigned-workouts]
      summary: Create assigned workout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignedWorkoutRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignedWorkout"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/my-assigned-workouts:
    get:
      tags: [student-assigned-workouts]
      summary: Get own assigned workouts
      responses:
        "200":
          description: Assigned workouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssignedWorkout"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/coach/templates:
    get:
      tags: [coach-templates]
      summary: List workout templates
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkoutTemplate"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [coach-templates]
      summary: Create workout template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutTemplate"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/coach/weekly-templates:
    get:
      tags: [coach-weekly-templates]
      summary: List weekly templates
      responses:
        "200":
          description: Weekly templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeeklyTemplate"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [coach-weekly-templates]
      summary: Create weekly template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyTemplate"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/coach/profile:
    put:
      tags: [coach-profile]
      summary: Update coach profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                coach_description:
                  type: string
                coach_public:
                  type: boolean
      responses:
        "200":
          description: Updated
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/coaches:
    get:
      tags: [coaches]
      summary: List public coaches
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by name or description
      responses:
        "200":
          description: Coaches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoachListItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/coach/achievements:
    get:
      tags: [achievements]
      summary: List own achievements
      responses:
        "200":
          description: Achievements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoachAchievement"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [achievements]
      summary: Create achievement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAchievementRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachAchievement"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/invitations:
    get:
      tags: [invitations]
      summary: List own invitations
      responses:
        "200":
          description: Invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invitation"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [invitations]
      summary: Create invitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/notifications:
    get:
      tags: [notifications]
      summary: List notifications
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/notifications/unread-count:
    get:
      tags: [notifications]
      summary: Get unread notification count
      responses:
        "200":
          description: Count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/notifications/read-all:
    put:
      tags: [notifications]
      summary: Mark all notifications as read
      responses:
        "200":
          description: All marked read
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/notification-preferences:
    get:
      tags: [notifications]
      summary: Get notification preferences
      responses:
        "200":
          description: Preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [notifications]
      summary: Update notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationPreferences"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/files:
    post:
      tags: [files]
      summary: Upload a file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/admin/stats:
    get:
      tags: [admin]
      summary: Platform statistics
      responses:
        "200":
          description: Stats
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/users:
    get:
      tags: [admin]
      summary: List all users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/admin/achievements/pending:
    get:
      tags: [admin]
      summary: List unverified achievements
      responses:
        "200":
          description: Pending achievements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoachAchievement"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    UserProfile:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        name:
          type: string
        avatar_url:
          type: string
        custom_avatar:
          type: string
          description: Base64-encoded custom avatar
        sex:
          type: string
          enum: [M, F, other]
        birth_date:
          type: string
          format: date
        age:
          type: integer
          description: Calculated from birth_date
        weight_kg:
          type: number
          format: float
        height_cm:
          type: integer
        language:
          type: string
          enum: [es, en]
        is_coach:
          type: boolean
        is_admin:
          type: boolean
        coach_description:
          type: string
        coach_public:
          type: boolean
        onboarding_completed:
          type: boolean
        has_coach:
          type: boolean
        coach_id:
          type: integer
          format: int64
        coach_name:
          type: string
        coach_avatar:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        sex:
          type: string
          enum: [M, F, other]
        birth_date:
          type: string
          format: date
        weight_kg:
          type: number
          format: float
        height_cm:
          type: integer
        language:
          type: string
          enum: [es, en]
        onboarding_completed:
          type: boolean
...
    CreateWorkoutRequest:
      type: object
      required: [date, type]
      properties:
        date:
          type: string
          format: date
        distance_km:
          type: number
          format: float
        duration_seconds:
          type: integer
        avg_pace:
          type: string
        calories:
          type: integer
        avg_heart_rate:
          type: integer
        feeling:
          type: integer
          minimum: 1
          maximum: 10
        type:
          type: string
          enum: [easy, tempo, intervals, long_run, race, fartlek, other]
        notes:
          type: string
        segments:
          type: array
          items:
            $ref: "#/components/schemas/SegmentRequest"
...